+++ src/rust/src/lib.rs
@@ -58,6 +58,7 @@ pub use single_cell::r_sc_data;
 pub use single_cell::r_sc_metacells;
 pub use single_cell::r_sc_plot_extraction;
 pub use single_cell::r_sc_processing;
+
 // meta cell
 pub use meta_cell::r_mc_processing;
 
@@ -74,20 +75,24 @@ extendr_module! {
     use r_stats;
     use r_svd_pca;
     use r_loess;
+
     // data
     use r_sparse;
     use r_synthetic;
     use r_h5;
+
     // enrichment
     use r_gsea;
     use r_gsva;
     use r_mitch;
     use r_oea;
+
     // graphs
     use r_page_rank;
     use r_snf;
     use r_graph_clustering;
     use r_knn;
+
     // methods
     use r_coremo;
     use r_dgrdl;
@@ -95,9 +100,11 @@ extendr_module! {
     use r_ica;
     use r_rbh;
     use r_cistarget;
+
     // ontology
     use r_go_elim;
     use r_similiarity;
+
     // single cell
     use r_count_obj;
     use r_sc_batch_corr;
@@ -105,4 +112,7 @@ extendr_module! {
     use r_sc_analysis;
     use r_sc_metacells;
     use r_sc_plot_extraction;
+
+    // meta cell
+    use r_mc_processing;
 }